+++ src/components/Home/Hero.jsx
@@ -1,4 +1,4 @@
-import ContentWrapper from "./ContentWrapper";
+import ContentWrapper from "../ContentWrapper";
 import { Link } from "react-router-dom";
 import { ArrowRight } from "lucide-react";
