+++ src/components/Home/RealResultsCaseStudy.jsx
@@ -1,4 +1,4 @@
-import ContentWrapper from "./ContentWrapper";
+import ContentWrapper from "../ContentWrapper";
 import { ArrowRight } from "lucide-react";
 
 const RealResultsCaseStudy = () => {
